selenium/requirements.txt
python-dateutil
pytest
selenium
